% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confrontation.R
\docType{methods}
\name{all,validation-method}
\alias{all,validation-method}
\title{Test if all validations resulted in TRUE}
\usage{
\S4method{all}{validation}(x, ..., na.rm = FALSE)
}
\arguments{
\item{x}{\code{validation} object (see \code{confront}).}

\item{...}{ignored}

\item{na.rm}{[\code{logical}] If \code{TRUE}, \code{NA} values
are removed before the result is computed.}
}
\description{
Test if all validations resulted in TRUE
}
\examples{
val <- check_that(women, height>60, weight>0)
all(val)
}
\seealso{
Other validation-methods: \code{\link{aggregate,validation-method}},
  \code{\link{any,validation-method}},
  \code{\link{barplot,validation-method}},
  \code{\link{check_that}}, \code{\link{compare}},
  \code{\link{confront}}, \code{\link{event}},
  \code{\link{plot,validation-method}},
  \code{\link{sort,validation-method}},
  \code{\link{summary}}, \code{\link{validation-class}},
  \code{\link{values}}
}
\concept{validation-methods}
