% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionset.R
\name{as.data.frame,expressionset-method}
\alias{as.data.frame,expressionset-method}
\title{Translate an expressionset to data.frame}
\usage{
\S4method{as.data.frame}{expressionset}(x, expand_assignments = TRUE, ...)
}
\arguments{
\item{x}{Object to coerce}

\item{expand_assignments}{Toggle substitution of `:=` assignments.}

\item{...}{arguments passed to other methods}
}
\value{
A \code{data.frame} with elements \code{rule}, \code{name},
 \code{label}, \code{origin}, \code{description}, and \code{created}.
}
\description{
Expressions are deparsed and combined in a \code{data.frame} with (some
of) their metadata. Observe that some information may be lost (e.g. options
local to the object).
}
\seealso{
Other expressionset-methods: 
\code{\link{as.data.frame}()},
\code{\link{created}()},
\code{\link{description}()},
\code{\link{label}()},
\code{\link{meta}()},
\code{\link{names<-,rule,character-method}},
\code{\link{origin}()},
\code{\link{plot,validator-method}},
\code{\link{summary}()},
\code{\link{variables}()},
\code{\link{voptions}()}
}
\concept{expressionset-methods}
