% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionset.R
\name{as.data.frame}
\alias{as.data.frame}
\title{Coerce to \code{data.frame}}
\usage{
as.data.frame(x, row.names = NULL, optional = FALSE, ...)
}
\arguments{
\item{x}{Object to coerce}

\item{row.names}{ignored}

\item{optional}{ignored}

\item{...}{arguments passed to other methods}
}
\description{
Coerce to \code{data.frame}
}
