\docType{methods}
\name{confront}
\alias{confront}
\alias{confront,indicator,data-method}
\alias{confront,validator,data-method}
\title{Confront data with a (set of) verifier(s)}
\usage{
confront(x, y, ...)

\S4method{confront}{indicator,data}(x, y, ...)

\S4method{confront}{validator,data}(x, y, ...)
}
\arguments{
  \item{x}{An R object carrying verifications}

  \item{y}{An R object carrying data}

  \item{...}{Arguments to be passed to other methods}
}
\description{
Confront data with a (set of) verifier(s)
}

