% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rule.R
\docType{methods}
\name{created<-,rule,POSIXct-method}
\alias{created<-,rule,POSIXct-method}
\title{Set creation timestamp}
\usage{
\S4method{created}{rule,POSIXct}(x) <- value
}
\arguments{
\item{x}{Object}

\item{value}{Value to set}
}
\description{
Set creation timestamp
}
\keyword{internal}

