% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\docType{methods}
\name{description<-,rule,character-method}
\alias{description<-,rule,character-method}
\title{Set description}
\usage{
\S4method{description}{rule,character}(x) <- value
}
\arguments{
\item{x}{Object}

\item{value}{Value to set}
}
\description{
Set description
}
\keyword{internal}
