% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confrontation.R
\docType{methods}
\name{errors}
\alias{errors}
\alias{errors,confrontation-method}
\alias{warnings,confrontation-method}
\title{Get messages from  a confrontation object}
\usage{
errors(x, ...)

\S4method{errors}{confrontation}(x, ...)

\S4method{warnings}{confrontation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{\link{confrontation}}}

\item{...}{Arguments to be passed to other methods.}
}
\description{
Get messages from  a confrontation object
}
\examples{

# create an error, by using a non-existent variable name
cf <- check_that(women, hite > 0, weight > 0)
# retrieve error messages
errors(cf)


}
\seealso{
\itemize{
 \item{\code{\link{confront}}}
}
}

