% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example.R
\name{Example}
\alias{Example}
\title{Component Title'}
\value{
a list value:
\itemize{
\item \code{example <- Example()}
\item \code{x <- 10}
\item \code{x |> Example()}
}
}
\description{
Component Description
}
