% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\docType{methods}
\name{expr}
\alias{expr}
\alias{expr,rule-method}
\title{Get expressions}
\usage{
expr(x, ...)

\S4method{expr}{rule}(x, ...)
}
\arguments{
\item{x}{Object}

\item{...}{options to be passed to other functions}
}
\description{
Get expressions
}
\keyword{internal}

