% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confrontation.R
\docType{class}
\name{indication-class}
\alias{indication}
\alias{indication-class}
\title{Store results of evaluating indicators}
\description{
\bold{This feature is currently experimental and may change in the future}
}
\section{Details}{

An \code{indication} stores a set of results generated by evaluating
an \code{\link{indicator}} in the context of data along with some metadata.
}

\section{Exported S4 methods for \code{indication}}{

\itemize{
 \item{Methods exported for objects of class \code{\link{confrontation}}}
 \item{\code{\link{summary,indication-method}}}
 \item{\code{\link{values,indication-method}}}
}
}

\section{See also}{

\itemize{
\item{\code{\link{confront}}}
\item{\code{\link{validation-class}}}
}
}
\keyword{internal}

