% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/indicator.R
\docType{class}
\name{indicator-class}
\alias{indicator-class}
\title{Store a set of rich indicator expressions}
\description{
\bold{This feature is currently experimental and may change in future versions}
}
\section{Details}{

An indicator stores a set of indicators. It is a child class of \code{\link{expressionset}} and
can be constructed with  \code{\link{indicator}}.
}

\section{Exported S4 methods for \code{validator}}{

  \itemize{
 \item{Methods inherited from \code{\link{expressionset}}}
 \item{\code{\link{confront}}}
 \item{\code{\link{compare}}}
}
}

\section{See also}{

\itemize{
 \item{\code{\link{expressionset}}}
}
}
\keyword{internal}

