% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{is_complete}
\alias{is_complete}
\alias{all_complete}
\title{Test for completeness of records}
\usage{
is_complete(...)

all_complete(...)
}
\arguments{
\item{...}{When used in a validation rule: a bare (unquoted) list of variable names.
When used directly, a comma-separated list of vectors of equal length.}
}
\value{
For \code{is_complete} A logical vector that is \code{FALSE} for each record
  that has a duplicate.

For \code{all_unique} a single \code{TRUE} or \code{FALSE}.
}
\description{
Utility function to make common tests easier.
}
\examples{
d <- data.frame(X = c('a','b',NA,'b'), Y = c(NA,'apple','banana','apple'), Z=1:4)
v <- validator(is_complete(X, Y))
values(confront(d, v))

}
\seealso{
Other cross-record-helpers: 
\code{\link{exists_any}()},
\code{\link{is_unique}()}
}
\concept{cross-record-helpers}
