\docType{methods}
\name{is_linear,validator-method}
\alias{is_linear}
\alias{is_linear,validator-method}
\title{Check for linear expressions}
\usage{
\S4method{is_linear}{validator}(x, ...)

is_linear(x, ...)
}
\arguments{
  \item{x}{An R object}

  \item{...}{Arguments to be passed to other methods.}
}
\value{
A \code{logical} vector
}
\description{
Check for linear expressions
}

