% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confrontation.R
\docType{methods}
\name{keyset}
\alias{keyset}
\alias{keyset,confrontation-method}
\title{Get key set stored with a confrontation}
\usage{
keyset(x)

\S4method{keyset}{confrontation}(x)
}
\arguments{
\item{x}{an object of class \code{confrontation}}
}
\value{
If a confrontation is created with the \code{key=} option
set, this function returns the key set, otherwise \code{NULL}
}
\description{
Get key set stored with a confrontation
}
\seealso{
Other confrontation-methods: \code{\link{[,expressionset-method}},
  \code{\link{as.data.frame,confrontation-method}},
  \code{\link{confrontation-class}},
  \code{\link{confront}}, \code{\link{errors}},
  \code{\link{event}},
  \code{\link{length,expressionset-method}},
  \code{\link{values}}
}
\concept{confrontation-methods}
