% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/expressionset.R
\docType{methods}
\name{label<-,expressionset,character-method}
\alias{label<-,expressionset,character-method}
\title{Set labels}
\usage{
\S4method{label}{expressionset,character}(x) <- value
}
\arguments{
\item{x}{Object}

\item{value}{Value to set}
}
\description{
Set labels
}
\examples{

# retrieve properties
v <- validator(turnover > 0, staff.costs>0)

# number of rules in v:
length(v)

# per-rule
created(v)
origin(v)
names(v)

# set properties
names(v)[1] <- "p1"

label(v)[1] <- "turnover positive"
description(v)[1] <- "
According to the official definition,
only positive values can be considered
valid turnovers.
"

# short description is also printed:
v

# print all info for first rule
v[[1]]


}

