% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rule.R
\docType{methods}
\name{label<-,rule,character-method}
\alias{label<-,rule,character-method}
\title{Set label}
\usage{
\S4method{label}{rule,character}(x) <- value
}
\arguments{
\item{x}{Object}

\item{value}{Value to set}
}
\description{
Set label
}
\keyword{internal}

