% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lumberjack.R
\docType{class}
\name{lbj_rules-class}
\alias{lbj_rules-class}
\alias{lbj_rules}
\title{Logging object to use with the lumberjack package}
\description{
Logging object to use with the lumberjack package
}
\section{Methods}{

\describe{
\item{\code{dump(file = "lbj_rules.csv", ...)}}{Dump logging info to csv file. 
All arguments in '...' except row.names are passed to 'write.csv'}

\item{\code{initialize(rules, verbose = TRUE)}}{Create object. Optionally toggle verbosity.}

\item{\code{log_data()}}{Return logged data as a data.frame}
}}

\seealso{
Other loggers: \code{\link{lbj_cells-class}}
}
