% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\name{match_cells}
\alias{match_cells}
\title{Create matching subsets of a sequence of data}
\usage{
match_cells(..., .list = NULL, id = NULL)
}
\arguments{
\item{...}{A sequence of \code{data.frame}s, possibly in the form of \code{<name>=<value>} pairs.}

\item{.list}{A list of \code{data.frame}s; will be concatenated with \code{...}.}

\item{id}{Names or indices of columns to use as index.}
}
\value{
A list of \code{data.frames}, subsetted and sorted so that all cells correspond.
}
\description{
Create matching subsets of a sequence of data
}
\seealso{
\itemize{
 \item{\code{\link{cells}}}
}
}
