% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retailers.R
\docType{data}
\name{nace_rev2}
\alias{nace_rev2}
\title{NACE classification code table}
\format{
A csv file, one NACE code per row.
}
\description{
Statistical Classification of Economic Activities. 



\itemize{ 
 \item Order \code{[integer]}
 \item Level \code{[integer]} NACE level
 \item Code  \code{[character]} NACE code
 \item Parent \code{[character]} parent code of \code{"Code"}
 \item Description \code{[character]}
 \item This_item_includes \code{[character]}
 \item This_item_also_includes \code{[character]}
 \item Rulings \code{[character]}
 \item This_item_excludes \code{[character]}
 \item Reference_to_ISIC_Rev._4 \code{[character]}
}
}
\seealso{
\code{\link{hierarchy}}

Other datasets: 
\code{\link{retailers}},
\code{\link{samplonomy}}
}
\concept{datasets}
