% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionset.R
\docType{methods}
\name{names,expressionset-method}
\alias{names,expressionset-method}
\title{Extract names}
\usage{
\S4method{names}{expressionset}(x)
}
\arguments{
\item{x}{An R object}
}
\value{
A \code{character} with names of rules occurring in \code{x}
}
\description{
Extract names
}
\examples{

# retrieve properties
v <- validator(turnover > 0, staff.costs>0)

# number of rules in v:
length(v)

# per-rule
created(v)
origin(v)
names(v)

# set properties
names(v)[1] <- "p1"

label(v)[1] <- "turnover positive"
description(v)[1] <- "
According to the official definition,
only positive values can be considered
valid turnovers.
"

# short description is also printed:
v

# print all info for first rule
v[[1]]


}

