% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rule.R
\docType{methods}
\name{names<-,rule,character-method}
\alias{names<-,rule,character-method}
\title{Set names}
\usage{
\S4method{names}{rule,character}(x) <- value
}
\arguments{
\item{x}{Object}

\item{value}{Value to set}
}
\description{
Set names
}
\keyword{internal}

