% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{number_missing}
\alias{number_missing}
\alias{fraction_missing}
\alias{row_missing}
\alias{col_missing}
\alias{number_unique}
\alias{any_missing}
\alias{any_duplicated}
\title{Missingness counters (DEPRECATED)}
\usage{
number_missing(...)

fraction_missing(...)

row_missing(...)

col_missing(...)

number_unique(...)

any_missing(...)

any_duplicated(...)
}
\arguments{
\item{...}{comma-separated list of variable names (not character). If no
variables are specified, the number of missings over all data is counted.}
}
\value{
For \code{number_missing}, the total number of missings over all specified variables.

For \code{fraction_missing}, the fraction of missings over all specified variables

For \code{row_missing} a vector with the number of missings per (sub)record defined by \code{...}.

For \code{col_missing} a vector with the number of missings per column 
   defined by \code{...}.

For \code{number_unique} the number of records, unique for 
  the columns specified in \code{...}.

For \code{any_missing}, \code{TRUE} if any \code{NA} occur in the columns
  specified in \code{...}.

For \code{any_duplicated}, \code{TRUE} if any (sub)records specified by
 \code{...} are duplicated, \code{FALSE} otherwise. Note that \code{NA} is matched with \code{NA}.
}
\description{
Missingness counters (DEPRECATED)
}
\keyword{internal}
