% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R
\docType{methods}
\name{origin<-,rule,character-method}
\alias{origin<-,rule,character-method}
\title{Set origin}
\usage{
\S4method{origin}{rule,character}(x) <- value
}
\arguments{
\item{x}{Object}

\item{value}{Value to set}
}
\description{
Set origin
}
\keyword{internal}

