\docType{methods}
\name{origin}
\alias{origin}
\alias{origin,verifier-method}
\title{Find out where expressions were defined}
\usage{
origin(x, ...)

\S4method{origin}{verifier}(x, ...)
}
\arguments{
  \item{x}{and R object}

  \item{...}{Arguments to be passed to other methods}
}
\value{
A \code{character} vector.
}
\description{
Find out where expressions were defined
}

