% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.R
\docType{methods}
\name{plot,validatorComparison-method}
\alias{plot,validatorComparison-method}
\title{Line graph of validatorComparison object}
\usage{
\S4method{plot}{validatorComparison}(x, xlab = "", ylab = "", las = 2,
  cex.axis = 0.8, cex.legend = 0.8, ...)
}
\arguments{
\item{x}{Object of class \code{validatorComparison}.}

\item{xlab}{[\code{character}] label for x axis (default none)}

\item{ylab}{[\code{character}] label for y axis (default none)}

\item{las}{[\code{numeric}] in \code{{0,1,2,3}} determining axis label rotation}

\item{cex.axis}{[\code{numeric}] Magnification with respect to the current
setting of \code{cex} for axis annotation.}

\item{cex.legend}{[\code{numeric}] Magnification with respect to the current
setting of \code{cex} for legend annotation and title.}

\item{...}{Graphical parameters, passed to \code{plot}. See \code{\link[graphics]{par}}.}
}
\description{
The performance of versions of a data set with regard to rule-based quality
requirements can be compared using using \code{\link{compare}}. The result is a
\code{validatorComparison} object. This method creates a line-graph, thus
suggesting an that an ordered sequence of data sets have been compared.  See
also \code{\link{barplot,validatorComparison-method}} for an unordered version.
}
\seealso{
Other comparing: \code{\link{as.data.frame,cellComparison-method}},
  \code{\link{as.data.frame,validatorComparison-method}},
  \code{\link{barplot,cellComparison-method}},
  \code{\link{barplot,validatorComparison-method}},
  \code{\link{cells}}, \code{\link{compare}},
  \code{\link{match_cells}},
  \code{\link{plot,cellComparison-method}}
}
