% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{+,indicator,indicator-method}
\alias{+,indicator,indicator-method}
\title{Combine two indicator objects}
\usage{
\S4method{+}{indicator,indicator}(e1, e2)
}
\arguments{
\item{e1}{a \code{\link{validator}}}

\item{e2}{a \code{\link{validator}}}
}
\description{
Combine two \code{\link{indicator}} objects by addition. A new \code{indicator} 
object is created with default (global) option values. Previously set options
are ignored.
}
\examples{
indicator(mean(x)) + indicator(x/median(x))
}
\concept{indicator-methods}
