% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validator.R
\docType{methods}
\name{+,validator,validator-method}
\alias{+,validator,validator-method}
\title{Combine two validator objects}
\usage{
\S4method{+}{validator,validator}(e1, e2)
}
\arguments{
\item{e1}{a \code{\link{validator}}}

\item{e2}{a \code{\link{validator}}}
}
\description{
Combine two \code{\link{validator}} objects by addition. A new \code{validator} 
object is created with default (global) option values. Previously set options
are ignored.
}
\section{Note}{

The \code{names} of the resulting object are made unique using \code{\link[base]{make.names}}.
}

\examples{
validator(x>0) + validator(x<=1)
}
\seealso{
Other validator-methods: \code{\link{plot,validator-method}},
  \code{\link{validator}}
}
\concept{validator-methods}
