% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retailers.R
\docType{data}
\name{retailers}
\alias{retailers}
\title{data on Dutch supermarkets}
\format{A csv file, one retailer per row.}
\description{
Anonymized and distorted data on revenue and cost structure for
  60 retailers. 
Currency is in thousands of Euros.

\itemize{
 \item size: Size class (0=undetermined)
 \item incl.prob: Probability of inclusion in the sample
 \item staff: Number of staff
 \item turnover: Amount of turnover
 \item other.rev: Amount of other revenue
 \item total.rev: Total revenue
 \item staff.costs: Number of staff employed
 \item total.costs: Total costs made
 \item profit: Amount of profit
 \item vat: Turnover reported for Value Added Tax
}
}
