% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rule.R
\docType{class}
\name{rule-class}
\alias{rule}
\alias{rule-class}
\title{A rich expression}
\description{
A rich expression
}
\section{Details}{

Technically, \code{rule} is a \code{call} object endowed with extra attributes such
as a name, a label and a description description, creation time and a reference to its origin.
Rule objects are not for direct use by users of the package, but may be of interest for
developers of this package, or packages depending on it.
}

\section{Exported S4 methods for \code{rule}}{

\itemize{
\item{\code{show}}
\item{\code{\link{origin}}}
\item{\code{\link{label}}}
\item{\code{\link{description}}}
\item{\code{\link{created}}}
}
}

\section{Private S4 methods for \code{rule}}{

\itemize{
 \item{validating}
 \item{linear}
 \item{expr}
 \item{is_tran_assign}
}
}

\section{See also}{

\itemize{
 \item{\code{\link{expressionset-class}}}
}
}
\keyword{internal}

