% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genericrules.R
\name{rx}
\alias{rx}
\alias{glob}
\title{Label objects for interpretation as pattern}
\usage{
rx(x)

glob(x)
}
\arguments{
\item{x}{Object to label as regular expression (\code{rx(x)}) or globbing
(\code{glob(x)}) pattern.}
}
\description{
Label objects  (typically strings or data frames containing keys combinations) 
to be interpreted as regular expression or globbing pattern.
}
