% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdmx.R
\name{sdmx_endpoint}
\alias{sdmx_endpoint}
\title{Get URL for known SDMX registry endpoints}
\usage{
sdmx_endpoint(registry = NULL)
}
\arguments{
\item{registry}{\code{[character]} name of the endpoint (case insensitve). If \code{registry}
is \code{NULL} (the default), the list of supported endpoints is returned.}
}
\description{
Convenience function storing URLs for SDMX endpoints.
}
\examples{
sdmx_endpoint()
sdmx_endpoint("ESTAT")
sdmx_endpoint("global")

}
\seealso{
Other sdmx: 
\code{\link{sdmx_codelist}()},
\code{\link{validator_from_dsd}()}
}
\concept{sdmx}
