\docType{methods}
\name{[,confrontation-method}
\alias{[,confrontation-method}
\alias{[,verifier-method}
\title{Select}
\usage{
\S4method{[}{confrontation}(x, i, j, ..., drop = TRUE)

\S4method{[}{verifier}(x, i, j, ..., drop = TRUE)
}
\arguments{
  \item{x}{An R object}

  \item{i}{an index (numeric, boolean, character)}

  \item{j}{not implemented}

  \item{drop}{not implemented}

  \item{...}{Arguments to be passed to other methods}
}
\value{
An new object, of the same class as \code{x} subsetted
according to \code{i}.
}
\description{
Select
}

