% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionset.R
\docType{methods}
\name{[[<-,expressionset-method}
\alias{[[<-,expressionset-method}
\title{Replace a rule in a ruleseta}
\usage{
\S4method{[[}{expressionset}(x, i) <- value
}
\arguments{
\item{x}{an R object}

\item{i}{index of length 1}

\item{value}{object of class \code{\link{rule}}}
}
\description{
Replace a rule in a ruleseta
}
\keyword{internal}
