% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionset.R
\name{[<-,expressionset-method}
\alias{[<-,expressionset-method}
\title{Replace a subset of an expressionset with another expressionset}
\usage{
\S4method{[}{expressionset}(x, i) <- value
}
\arguments{
\item{x}{an R object inheriting from \code{expressionset}}

\item{i}{a \code{logical}, \code{character}, or \code{numeric} index}

\item{value}{an R object of the same class as \code{x}}
}
\description{
Replace a subset of an expressionset with another expressionset
}
\keyword{internal}
