% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse.R, R/expressionset.R
\name{.PKGOPT}
\alias{.PKGOPT}
\alias{.ini_expressionset_cli}
\alias{.ini_expressionset_df}
\alias{.ini_expressionset_yml}
\alias{.show_expressionset}
\alias{.get_exprs}
\alias{.blocks_expressionset}
\title{Services for extending 'validate'}
\usage{
.PKGOPT(..., .__defaults = FALSE, .__reset = FALSE)

.ini_expressionset_cli(obj, ..., .prefix = "R")

.ini_expressionset_df(obj, dat, .prefix = "R")

.ini_expressionset_yml(obj, file, .prefix = "R")

.show_expressionset(obj)

.get_exprs(
  x,
  ...,
  expand_assignments = FALSE,
  expand_groups = TRUE,
  vectorize = TRUE,
  replace_dollar = TRUE,
  replace_in = TRUE,
  lin_eq_eps = x$options("lin.eq.eps"),
  lin_ineq_eps = x$options("lin.ineq.eps"),
  dat = NULL
)

.blocks_expressionset(x)
}
\arguments{
\item{...}{Comma-separated list of expressions}

\item{.__defaults}{toggle default options}

\item{.__reset}{togle reset options}

\item{obj}{an expressionset object}

\item{.prefix}{Prefix to use in default names.}

\item{dat}{Optionally, a \code{data.frame} containing the data to which the
expressions will be applied. When provided, the only equalities \code{A==B}
that will be translated to \code{abs(A-B)<lin.eq.eps} are those where all 
occurring variables are numeric in \code{dat}.}

\item{file}{a filename}

\item{x}{An expressionset object}

\item{expand_assignments}{Substitute assignments?}

\item{expand_groups}{Expand groups?}

\item{vectorize}{Vectorize if-statements?}

\item{replace_dollar}{Replace dollar with bracket index?}

\item{varlist:}{a character vector of variables to search through 
when groups are defined with regexps.}
}
\description{
Functions exported silently to allow for cross-package inheritance 
of the \code{\link{expressionset}} object. These functions are never
needed in scripts or statistical production code.
}
\section{Details}{

This function is aimed at developers importing the package and 
not at direct users of \pkg{validate}.
}

\keyword{internal}
