% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confrontation.R
\docType{class}
\name{validation-class}
\alias{validation}
\alias{validation-class}
\title{Store results of evaluating validating expressions}
\description{
Store results of evaluating validating expressions
}
\section{Details}{

A object of class \code{validation} stores a set of results generated by
evaluating an \code{\link{validator}} in the context of data along with some
metadata.
}

\section{Exported S4 methods for \code{validation}}{

\itemize{
 \item{Methods exported for objects of class \code{\link{confrontation}}}
 \item{\code{\link{summary,validation-method}}}
 \item{\code{\link{values,validation-method}}}
 \item{\code{\link{barplot,validation-method}}}
 \item{\code{\link{aggregate,validation-method}}}
 \item{\code{\link{sort,validation-method}}}
}
}

\section{See also}{

\itemize{
\item{\code{\link{confront}}}
\item{\code{\link{validator}}}
}
}

