% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validator.R
\name{Validator}
\alias{Validator}
\title{Validator'}
\value{
a list value:
\itemize{
\item \code{input |> validate[['Is.Not.NULL']](argument)}
\item \code{input |> validate[['Is.Data.Frame']]()}
\item \code{input |> validate[['Is.Character']]()}
\item \code{input |> validate[['Is.Empty']]()}
\item \code{input |> validate[['Is.Not.Empty']]()}
\item \code{input |> validate[['Has.One.Row']]()}
\item \code{input |> validate[['Is.UUID']](identifier)}
}
}
\description{
Validator which returns a generic set of validators and accompanied exceptions.
}
