\docType{methods}
\name{values}
\alias{impact}
\alias{impact,validatorValue-method}
\alias{severity}
\alias{severity,}
\alias{severity,validatorValue-method}
\alias{values}
\alias{values,confrontation-method}
\alias{values,validatorValue-method}
\title{Get values from object}
\usage{
values(x, ...)

\S4method{values}{confrontation}(x, ...)

\S4method{values}{validatorValue}(x, simplify = TRUE, ...)

severity(x, ...)

\S4method{severity}{validatorValue}(x, ...)

impact(x, ...)

\S4method{impact}{validatorValue}(x, ...)
}
\arguments{
  \item{x}{an R object}

  \item{...}{Arguments to pass to or from other methods}
}
\description{
Get values from object

Get severity values

Get impact values
}

