\docType{methods}
\name{variables}
\alias{variables}
\alias{variables,verifier-method}
\title{Extract variable names}
\usage{
variables(x, ...)

\S4method{variables}{verifier}(x, ...)
}
\arguments{
  \item{x}{An R object}

  \item{...}{Methods to be passed to other methods.}
}
\description{
Extract variable names
}

