% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confrontation.R
\docType{class}
\name{confrontation-class}
\alias{confrontation}
\alias{confrontation-class}
\title{Superclass storing results of confronting data with rules}
\description{
Superclass storing results of confronting data with rules
}
\section{Details}{

This class is aimed at developers of this package or packages depending on 
it. It is the parent of classes \code{\link{indication}} and 
\code{\link{validation}} which are user-facing.

Using \code{\link{confront}}, a set of rules can be executed in the context
of one or more (nested) environments holding data. The results of such evaluations
are stored in a \code{confrontation} object along with metadata. 

We strongly advise against accessing the data fields or methods internal to
this object directly, as we may change or remove them without notice. Use
the exported methods listed below in stead.
}

\section{Exported S4 methods for \code{confrontation}}{


\itemize{
 \item{\code{show}}
 \item{\code{\link{values}}}
 \item{\code{\link{warnings,confrontation-method}}}
 \item{\code{\link{errors}}}
 \item{\code{\link{aggregate}}}
 \item{\code{\link{sort}}}
 \item{\code{\link{length,confrontation-method}}}
 \item{\code{\link{[,confrontation-method}}}
}
}

\section{Private S4 methods for \code{confrontation}}{


Currently N/A
}

\section{See also}{

\itemize{
 \item{\code{\link{validation}}}
 \item{\code{\link{indication}}}
 \item{\code{\link{expressionset}}}
}
}
\keyword{internal}

