% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rule.R, R/expressionset.R
\name{created}
\alias{created}
\alias{created<-}
\alias{created,rule-method}
\alias{created<-,rule,POSIXct-method}
\alias{created,expressionset-method}
\alias{created<-,expressionset,POSIXct-method}
\title{Creation timestamp}
\usage{
created(x, ...)

created(x) <- value

\S4method{created}{rule}(x, ...)

\S4method{created}{rule,POSIXct}(x) <- value

\S4method{created}{expressionset}(x, ...)

\S4method{created}{expressionset,POSIXct}(x) <- value
}
\arguments{
\item{x}{and R object}

\item{...}{Arguments to be passed to other methods}

\item{value}{Value to set}
}
\value{
A \code{POSIXct} vector.
}
\description{
Creation timestamp
}
\examples{

# retrieve properties
v <- validator(turnover > 0, staff.costs>0)

# number of rules in v:
length(v)

# per-rule
created(v)
origin(v)
names(v)

# set properties
names(v)[1] <- "p1"

label(v)[1] <- "turnover positive"
description(v)[1] <- "
According to the official definition,
only positive values can be considered
valid turnovers.
"

# short description is also printed:
v

# print all info for first rule
v[[1]]



# retrieve properties
v <- validator(turnover > 0, staff.costs>0)

# number of rules in v:
length(v)

# per-rule
created(v)
origin(v)
names(v)

# set properties
names(v)[1] <- "p1"

label(v)[1] <- "turnover positive"
description(v)[1] <- "
According to the official definition,
only positive values can be considered
valid turnovers.
"

# short description is also printed:
v

# print all info for first rule
v[[1]]


}
\seealso{
Other expressionset-methods: 
\code{\link{as.data.frame,expressionset-method}},
\code{\link{as.data.frame}()},
\code{\link{description}()},
\code{\link{label}()},
\code{\link{meta}()},
\code{\link{names<-,rule,character-method}},
\code{\link{origin}()},
\code{\link{plot,validator-method}},
\code{\link{summary}()},
\code{\link{variables}()},
\code{\link{voptions}()}
}
\concept{expressionset-methods}
