% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/indicator.R
\name{indicator}
\alias{indicator}
\title{Define indicators for data}
\usage{
indicator(..., .file, .data)
}
\arguments{
\item{...}{A comma-separated list of indicator definitions}

\item{.file}{(optional) A character vector of file locations}
}
\description{
\bold{This feature is currently experimental and may change in future versions}
}
\details{
\code{\link{indicator}}
}
\examples{
# create an indicator for the number of missing x in data set


I <- indicator( 
 sum(is.na(.))               # number of missing variables
 , sum(is.na(.[c("x","y")])) # number of missing x and y
 , mean(is.na(.))            # fraction of missing variables
 , sum(x)
 , mean(x)
) 

dat <- data.frame(x=1:2, y=c(NA,1))
C <- confront(dat, I)
values(C)
}
\seealso{
\code{\link{syntax}}
}
\keyword{internal}
