% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syntax.R
\name{is_unique}
\alias{is_unique}
\alias{all_unique}
\title{Test for uniquenes of records}
\usage{
is_unique(...)

all_unique(...)
}
\arguments{
\item{...}{When used in a validation rule: a bare (unquoted) list of variable names.
When used directly, a comma-separated list of vectors of equal length.}
}
\value{
For \code{is_unique} A logical vector that is \code{FALSE} for each record
  that has a duplicate.

For \code{all_unique} a single \code{TRUE} or \code{FALSE}.
}
\description{
Utility function to make common tests easier.
}
\examples{
# check that height-weight combinations are unique
d <- data.frame(X = c('a','b','c','b'), Y = c('banana','apple','banana','apple'), Z=1:4)
v <- validator(is_unique(X, Y))
values(confront(d, v))

}
