% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionset.R, R/confrontation.R
\docType{methods}
\name{length,expressionset-method}
\alias{length,expressionset-method}
\alias{length,confrontation-method}
\title{Determine the number of elements in an object.}
\usage{
\S4method{length}{expressionset}(x)

\S4method{length}{confrontation}(x)
}
\arguments{
\item{x}{An R object}
}
\description{
Determine the number of elements in an object.
}
