% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confrontation.R, R/expressionset.R
\docType{methods}
\name{length,confrontation-method}
\alias{length}
\alias{length,confrontation-method}
\alias{length,expressionset-method}
\alias{validate-length}
\title{Get object lenght}
\usage{
\S4method{length}{confrontation}(x)

length(x)

\S4method{length}{expressionset}(x)
}
\arguments{
\item{x}{An R object}
}
\description{
Get object lenght
}
\examples{

# retrieve properties
v <- validator(turnover > 0, staff.costs>0)

# number of rules in v:
length(v)

# per-rule
created(v)
origin(v)
names(v)

# set properties
names(v)[1] <- "p1"

label(v)[1] <- "turnover positive"
description(v)[1] <- "
According to the official definition,
only positive values can be considered
valid turnovers.
"

# short description is also printed:
v

# print all info for first rule
v[[1]]


}
\seealso{
\itemize{
 \item{\code{\link{expressionset}}}
 \item{\code{\link{confrontation}}}
}
}

