% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/confrontation.R, R/expressionset.R
\docType{methods}
\name{[,confrontation,ANY,ANY-method}
\alias{[,confrontation,ANY,ANY-method}
\alias{[,confrontation-method}
\alias{[,expressionset,ANY,ANY-method}
\alias{[,expressionset-method}
\alias{[[,expressionset-method}
\title{Select a subset}
\usage{
\S4method{[}{confrontation,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[}{expressionset,ANY,ANY}(x, i, j, ..., drop = TRUE)

\S4method{[[}{expressionset}(x, i, j, ..., exact = TRUE)
}
\arguments{
\item{x}{An R object}

\item{i}{an index (numeric, boolean, character)}

\item{j}{not implemented}

\item{...}{Arguments to be passed to other methods}

\item{drop}{not implemented}

\item{exact}{Not implemented}
}
\value{
An new object, of the same class as \code{x} subsetted according to \code{i}.
}
\description{
Select a subset
}
\section{Details}{

The \code{options} attribute will be cloned
}

