% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expressionset.R, R/confrontation.R
\docType{methods}
\name{[,expressionset-method}
\alias{[,expressionset-method}
\alias{[[,expressionset-method}
\alias{[,confrontation-method}
\title{Select a subset}
\usage{
\S4method{[}{expressionset}(x, i, j, ..., drop = TRUE)

\S4method{[[}{expressionset}(x, i, j, ..., exact = TRUE)

\S4method{[}{confrontation}(x, i, j, ..., drop = TRUE)
}
\arguments{
\item{x}{An R object}

\item{i}{an index (numeric, boolean, character)}

\item{j}{not implemented}

\item{...}{Arguments to be passed to other methods}

\item{drop}{not implemented}

\item{exact}{Not implemented}
}
\value{
An new object, of the same class as \code{x} subsetted according to \code{i}.
}
\description{
Select a subset
}
\section{Details}{

The \code{options} attribute will be cloned
}

