% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confrontation.R
\docType{class}
\name{validation-class}
\alias{validation-class}
\alias{validation}
\title{Store results of evaluating validating expressions}
\description{
Store results of evaluating validating expressions
}

\section{Details}{

A object of class \code{validation} stores a set of results generated by
evaluating an \code{\link{validator}} in the context of data along with some
metadata.
}

\seealso{
Other validation-methods: 
\code{\link{aggregate,validation-method}},
\code{\link{all,validation-method}},
\code{\link{any,validation-method}},
\code{\link{barplot,validation-method}},
\code{\link{check_that}()},
\code{\link{compare}()},
\code{\link{confront}()},
\code{\link{event}()},
\code{\link{names<-,rule,character-method}},
\code{\link{plot,validation-method}},
\code{\link{sort,validation-method}},
\code{\link{summary}()},
\code{\link{values}()}
}
\concept{validation-methods}
