% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validator.R
\name{Validator}
\alias{Validator}
\title{Validator'}
\value{
a list value:
\itemize{
\item \code{input |> validate[['is.not.NULL']](argument)}
\item \code{input |> validate[['is.data.frame']]()}
\item \code{input |> validate[['is.character']]()}
\item \code{input |> validate[['is.empty']]()}
\item \code{input |> validate[['is.not.empty']]()}
\item \code{input |> validate[['has.one.row']]()}
\item \code{input |> validate[['is.UUID']](identifier)}
}
}
\description{
Validator which returns a generic set of validators and accompanied exceptions.
}
