% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Validator.R
\name{Validator}
\alias{Validator}
\title{Validator'}
\value{
a list value:
\itemize{
\item \code{input |> validate[['is.not.NULL']](argument)}
\item \code{input |> validate[['is.data.frame']]()}
\item \code{input |> validate[['is.vector']]()}
\item \code{input |> validate[['is.character']]()}
\item \code{input |> validate[['is.numeric']]()}
\item \code{input |> validate[['is.integer']]()}
\item \code{input |> validate[['is.double']]()}
\item \code{input |> validate[['is.logical']]()}
\item \code{input |> validate[['is.complex']]()}
\item \code{input |> validate[['is.raw']]()}
\item \code{input |> validate[['is.factor']]()}
\item \code{input |> validate[['is.matrix']]()}
\item \code{input |> validate[['is.array']]()}
\item \code{input |> validate[['is.list']]()}
\item \code{input |> validate[['is.empty']]()}
\item \code{input |> validate[['is.not.empty']]()}
\item \code{input |> validate[['has.one.row']]()}
\item \code{input |> validate[['has.length']](length)}
\item \code{input |> validate[['has.name']](name)}
\item \code{input |> validate[['is.UUID']]()}
\item \code{input |> validate[['is.email']]()}
\item \code{input |> validate[['is.greater.than']](value)}
\item \code{input |> validate[['is.less.than']](value)}
\item \code{input |> validate[['is.equal.to']](value)}
\item \code{input |> validate[['is.one.of']](values)}
}
}
\description{
Validator which returns a generic set of validators and accompanied exceptions.
}
