% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confrontation.R
\docType{methods}
\name{values}
\alias{values}
\alias{values,confrontation-method}
\alias{values,validation-method}
\alias{values,indication-method}
\title{Get values from object}
\usage{
values(x, ...)

\S4method{values}{confrontation}(x, ...)

\S4method{values}{validation}(x, simplify = TRUE, drop = TRUE, ...)

\S4method{values}{indication}(x, simplify = TRUE, drop = TRUE, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{Arguments to pass to or from other methods}

\item{simplify}{Combine results with similar dimension structure into arrays?}

\item{drop}{if a single vector or array results, drop 'list' attribute?}
}
\description{
Get values from object
}
