% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confrontation.R
\name{values}
\alias{values}
\alias{values,confrontation-method}
\alias{values,validation-method}
\alias{values,indication-method}
\title{Get values from object}
\usage{
values(x, ...)

\S4method{values}{confrontation}(x, ...)

\S4method{values}{validation}(x, simplify = TRUE, drop = TRUE, ...)

\S4method{values}{indication}(x, simplify = TRUE, drop = TRUE, ...)
}
\arguments{
\item{x}{an R object}

\item{...}{Arguments to pass to or from other methods}

\item{simplify}{Combine results with similar dimension structure into arrays?}

\item{drop}{if a single vector or array results, drop 'list' attribute?}
}
\description{
Get values from object
}
\seealso{
Other confrontation-methods: 
\code{\link{[,expressionset-method}},
\code{\link{as.data.frame,confrontation-method}},
\code{\link{confront}()},
\code{\link{confrontation-class}},
\code{\link{errors}()},
\code{\link{event}()},
\code{\link{keyset}()},
\code{\link{length,expressionset-method}}

Other validation-methods: 
\code{\link{aggregate,validation-method}},
\code{\link{all,validation-method}},
\code{\link{any,validation-method}},
\code{\link{barplot,validation-method}},
\code{\link{check_that}()},
\code{\link{compare}()},
\code{\link{confront}()},
\code{\link{event}()},
\code{\link{names<-,rule,character-method}},
\code{\link{plot,validation-method}},
\code{\link{sort,validation-method}},
\code{\link{summary}()},
\code{\link{validation-class}}
}
\concept{confrontation-methods}
\concept{validation-methods}
